Support compilation benchmark
